requirements.txt
pdf2image>=3.1.0
Pillow>=10.0.0
pytesseract>=0.3.10
watchdog>=3.0.0
PyPDF2>=3.0.0
